third_party/blink/web_tests/external/wpt/html/browsers/sandboxing/sandbox-document-open-mutation.window-expected.txt

This is a testharness.js-based test.
FAIL Using document.open() against a document from a different window must not mutate the other window's sandbox flags assert_equals: expected "not sandboxed" but got "sandboxed"
Harness: the test ran to completion.
